nsc-typechecker: move instruction typechecking in a submodule
